Fix: Product.price int → Long 변경 및 OrderItem 상품명·가격 하드코딩 제거 #71

+++ src/main/java/com/bootcamp/paymentdemo/product/Product.java
@@ -20,7 +20,7 @@ public class Product extends BaseEntity {
 
     @Column(nullable = false)
     @Min(0)
-    private int price;
+    private Long price;  // int → Long으로 변경 (OrderItem.price 타입과 통일, 형변환 불필요)
 
     @Column(nullable = false)
     @Min(0)
@@ -31,7 +31,7 @@ public class Product extends BaseEntity {
     @Enumerated(value = EnumType.STRING)
     private ProductCategory productCategory;
 
-    public Product(String name, int price, int stock, String url, ProductCategory productCategory) {
+    public Product(String name, Long price, int stock, String url, ProductCategory productCategory) {
         this.name = name;
         this.price = price;
         this.stock = stock;
